Crypto_Tracker/requirements.txt
streamlit
pandas
plotly
yfinance
yfinance
requests
PyGithub
python-dotenv